fix: move AccAddrFromBech32 function to common package and add constant for Stride bech32
